Create HaircutsProject
This project stores a list of haircut appointments with details such as name, time, phone number, and gender. When the page loads, it displays the phone number of the first person in the list. It demonstrates how to work with arrays of objects in JavaScript and update the HTML content dynamically using window.onload.